### 2019-04-20 43a671b
Updated paper.md and paper.bib for joss submission after seeing paper rendering
