+++ paper.bib
@@ -8,13 +8,13 @@
 
 @online{ppiclF,
   author = {Zwick, D.},
-  title = {ppiclF.},
+  title = {{ppiclF.}},
   year = 2019,
   url = {https://github.com/dpzwick/ppiclF},
   urldate = {2019-04-19}
 }
 
-@online{ppiclF-doc,
+@online{ppiclf-doc,
   author = {Zwick, D.},
   title = {{A Parallel Particle-In-Cell Library in Fortran (ppiclF).}},
   year = 2019,
