### 2019-04-20 #1396 @jedbrown
@dpzwick Thanks for your submission. The paper is on the short side for JOSS and I think would benefit from including comparison with related method/software (some of which may be applications rather than libraries) and perhaps something on performance (if that is meant to be a distinguishing feature). The paper is intended to give a prospective user the information needed to make choices that match their research needs. Reviewers will also ask about [community guidelines ](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) so you can get a jump on it by adding that while we assemble reviewers.
