### 2019-04-21 72ae164
Updated paper.md and paper.bib per JOSS suggestion
