+++ paper.bib
@@ -11,7 +11,7 @@
   title = {{ppiclF.}},
   year = 2019,
   url = {https://github.com/dpzwick/ppiclF},
-  urldate = {2019-04-19}
+  urldate = {2019-04-21}
 }
 
 @online{ppiclf-doc,
@@ -19,5 +19,21 @@
   title = {{A Parallel Particle-In-Cell Library in Fortran (ppiclF).}},
   year = 2019,
   url = {https://dpzwick.github.io/ppiclF-doc},
-  urldate = {2019-04-19}
+  urldate = {2019-04-21}
+}
+
+@online{nek5000,
+  title = {{NEK5000.}},
+  year = 2019,
+  organization = {Argonne National Laboratory, Illinois},
+  url = {https://nek5000.mcs.anl.gov},
+  urldate = {2019-04-21}
+}
+
+@online{cmt-nek,
+  title = {{CMT-nek.}},
+  year = 2019,
+  organization = {The Center for Compressible Multiphase Turbulence at the University of Florida},
+  url = {https://github.com/UFCCMT/Nek5000},
+  urldate = {2019-04-21}
 }
