### 2019-04-21 #1396 @dpzwick
Thanks for the suggestions @jedbrown. I updated the paper.
