### 2019-04-23 #1400 @markadams
I think having the data lurk in common blocks will limit the extensibility of this project. I would suggest a more object oriented approach like provide a particle list object and iterators of some sort, and methods that operate on these lists. For instance, using multiple species of particles does not look feasible now. I don't know what the long term goal of this project is but it is young and I am suggesting a major refactoring -- the longer you wait the harder it will be to do.
