### 2019-04-23 #1400 @dpzwick
Thank you, that is a good suggestion. Indeed I think we may head that direction in the future. Either by upgrading to Fortran 90 or possibly even a while C++ overhaul.

While this would be helpful, it shouldn't hinder the current capabilities of ppiclF. Multiple species can be done already since the user can solve whichever equations they want and also store arbitrary numbers of properties in their simulation.

I will keep this in mind for future versions though.
