### 2019-04-27 dpzwick/ppiclF#25 @iljah
PPICLF_USER.h: No such file or directory
Trying to compile stokes 2d c++ example in `.../ppiclF/examples/stokes_2d/c++` but getting:
```
.../mpic++ -z muldefs -c test.cpp -I.../ppiclF/source
In file included from test.cpp:5:
.../ppiclF/source/PPICLC.h:1:10: fatal error: PPICLF_USER.h: No such file or directory
 #include "PPICLF_USER.h"
          ^~~~~~~~~~~~~~~
compilation terminated.
```
Adding ` -I ../user_routines` to CFLAGS seemed to fix this.
