### 2019-04-27 #1400 @iljah
Documentation (README.md as far as I can tell) could use a bit longer statement of need and is missing installation instructions and community guidelines.
