### 2019-04-27 #1400 @iljah
According to my instructions there should be API documentation but I don't see any in the repository. Where can I learn e.g. how `ppiclc_solve_InitBoxFilter` works?
