### 2019-04-27 #1400 @iljah
Readme also doesn't say anything about testing the library.
