### 2019-04-27 dpzwick/ppiclF#26 @iljah
/usr/bin/ld: cannot find -lppiclF
When compiling stokes 2d test I get:
```
...
***************************
*** USER OBJECT SUCCESS ***
***************************

.../mpif77 -z muldefs -I ../user_routines -o test.out test.o -L.../ppiclF/source -lppiclF -lstdc++
/usr/bin/ld: cannot find -lppiclF
collect2: error: ld returned 1 exit status
make: *** [Makefile:26: linkObjs] Error 1
```
Looks like library should be added as dependency so that it's compiled automatically. Or user could also be told to install/compile the library first...
