### 2019-04-27 dpzwick/ppiclF#27 @iljah
gslib uses hard-coded mpi c++ compiler
I have several MPI compilers installed and none are in default path so installation of gslib fails:
```
gslib]$ ./install
+ make -j4 DESTDIR=.. ADDUS= USREXIT=0 BLAS=2 PREFIX=gslib_ FPREFIX=fgslib_
mpicc -O2 -I./src  -DMPI -DUNDERSCORE -DGLOBAL_LONG_LONG -DPREFIX=gslib_ -DFPREFIX=fgslib_ -c src/gs.c -o src/gs.o
make: mpicc: Komentoa ei löytynyt
make: *** [Makefile:130: src/gs.o] Virhe 127
```
Is it possible to customize mpi c++ compiler path when installing gslib through ppiclF?
