### 2019-04-27 #1400 @iljah
Added some installation issues to library repo:
 - https://github.com/dpzwick/ppiclF/issues/25
 - https://github.com/dpzwick/ppiclF/issues/26
 - https://github.com/dpzwick/ppiclF/issues/27
