### 2019-04-27 dpzwick/ppiclF#28 @iljah
No rule to make target 'source/ppiclf_user.f'
Trying to compile the library gives:
```
$ make
cd ./3rd_party/gslib; \
./install

***********************
*** INSTALLED GSLIB ***
***********************

make: *** No rule to make target 'source/ppiclf_user.f', needed by 'getObjs'. Seis.
```
Am I supposed to create ppiclf_user.f?
