### 2019-04-27 ecf2aa0
Update README.md
Updated with suggestions from JOSS
