+++ README.md
@@ -2,28 +2,22 @@
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/dpzwick/ppiclF/blob/master/LICENSE)
 
 # ppiclF
-A parallel particle-in-cell library in Fortran.
+ppiclF is a parallel particle-in-cell library written in Fortran. Its main purpose is to provide a unified and scalable interface for a user to solve the following system of differential equations
 
-* Integration for the system:
            
            d y
-           ---  =  ydot, 
+           ---  =  ydot.
            d t
-           
-  where y and ydot are vectors that are entirely
-  user defined.
-       
-* Open MPI parallelization allows billions of equations
-  to be solved.
-       
-* Load balances equations based on spatial position of
-  particles.
 
-* Links with both Fortran and C++ external code.
-       
-* Allows simple user input of external overlapping mesh
-  for interactions between particles and external mesh,
-  including interpolation and projection.
-       
-* Includes optional fast binned parallel nearest neighbor
-  search between particles within a user defined distance.
+
+* See [documentation website](https://dpzwick.github.io/ppiclF-doc) for more details, theory, examples, questions, etc.
+
+* On-the-fly load-balancing of the system of equations across MPI processing ranks based on the coordinates associated with each particle. 
+
+* Simple user input of an external overlapping mesh for interactions between particles and their nearby cells.
+
+* Optional fast binned parallel nearest neighbor search between particles within a user specified distance so that more sophisticated user-implemented right-hand-side forcing models can easily be evaluated. 
+
+* Algorithms have demonstrated scalability to 100,000 processors, allowing billions of equations to be solved simultaneously. 
+
+* Links to both Fortran and C++ external code as a library.
