### 2019-04-27 e906cc9
Allowed gslib to use the same fortran and C compilers as main Makefile
Currently, only mpif77, mpif90, and mpicc are the only supported
compilers.

Fixes: https://github.com/dpzwick/ppiclF/issues/27
