### 2019-04-27 dpzwick/ppiclF#27 @dpzwick
Thanks for the feedback @iljah. I made the gslib install with the same compilers as the main Makefile so it isn't hard-coded anymore. Although currently ppiclF is only support mpicc (C compiler) and mpif77/mpif90 (Fortran comilers).
