### 2019-04-27 dpzwick/ppiclF#25 @dpzwick
Hi @iljah. Thanks for the feedback. I think there must be a misunderstanding when compiling. I would follow the quickstart guide on the documentation website (https://dpzwick.github.io/ppiclF-doc/quickstart.html). ppiclF must be built for each individual case as it says on the website. So the workflow is: 1) copy ppiclf_user.f and PPICLF_USER.h from an example into the source directory and then 2) make.
