### 2019-04-27 #1400 @dpzwick
Thank you for the feedback @iljah. I think issues dpzwick/ppiclF#25, dpzwick/ppiclF#26, and dpzwick/ppiclF#28 are a simple misunderstanding of not being able to see the documentation website.

While the link to the website was in the compiled JOSS pdf, I made it more clear in the README.md file along with your previous suggestions to that file (see [this commit](https://github.com/dpzwick/ppiclF/commit/ecf2aa00fffe62100f66631959353c94195489f4)).

As for issue dpzwick/ppiclF#27, I agree. It was confusing that even when a different compiler is specified gslib would compile with a hard-coded mpicc. I updated the source so gslib will use the same compiler as the main ppiclF Makefile. Although I also note that the only supported compilers currently are mpicc, mpif77 and mpif90.
