### 2019-04-27 #1400 @dpzwick
Also, to address your question above about where to find information about items such as: ppiclc_solve_InitBoxFilter, there is extensive information on the documentation website [here](https://dpzwick.github.io/ppiclF-doc/user/external.html).
