### 2019-04-28 dpzwick/ppiclF#27 @iljah
Thanks but mpicc isn't always the name of mpi c compiler. For example Cray machines have had something like cc as mpi c, CC as mpi c++ and ftn as fortran compiler name for decades. GNU autotools also allow folks to specify executable paths and names with environment variables. Please allow folks to specify full path to compilers.

Also, mpicc is just a wrapper so you can't make any assumptions about what compiler is run by executing mpicc, thus only allowing mpicc as executable name doesn't make sense to me.
