### 2019-04-28 cbf03ff
Compiler specification
The full path is allowed to be specified in the main Makefile to valid Fortran and C MPI compilers.

Fixes: https://github.com/dpzwick/ppiclF/issues/27
