### 2019-04-29 dpzwick/ppiclF#25 @iljah
Instructions at https://dpzwick.github.io/ppiclF-doc/quickstart/build.html don't work:
```
$ cp examples/stokes_2d/user_source/ppiclf_user.f source/
cp: cannot stat 'examples/stokes_2d/user_source/ppiclf_user.f': No such file or directory
$ cp examples/stokes_2d/user_source/PPICLF_USER.h source/
cp: cannot stat 'examples/stokes_2d/user_source/PPICLF_USER.h': No such file or directory
```
