### 2019-04-29 #1400 @iljah
Documentation on external website should be ok now that there's also a link to it in readme but it's also missing a statement of need or at least the target audience.

About performance: is there are particular reason why ghost particle send doesn't seem to scale past 1000 ranks, or at all? Also how many particles / s does ppiclF process on one core?

A link to vulcan page at LLNL in performance section would be nice.
