### 2019-04-29 #1400 @dpzwick
Thank you for the feedback. I updated the readme to include a target audience in [this commit](https://github.com/dpzwick/ppiclF/commit/db29738bc62cc4ae659a27a4ed240f668a736f97) and [this commit](https://github.com/dpzwick/ppiclF/commit/0f154f3cd1c2b5ee8f28e0cde1ad4f45b3921b8a).

Regarding the ghost particles, the reason for this is the underlying all-to-all communication. At best all-to-all communication can be log_2 (R) where R is number of MPI ranks unless some very specific refactoring of communication patterns is done. In the weak scaling, the particles per rank is fixed so it is expected to see this log scaling.

Regarding the number of particles on a single core, the number is only limited by the available memory per core. In the performance section it was varied between 256 - 24,576 when the strong scaling was done, although it could have been higher if needed.

I also added a link to Vulcan on the website in [this commit](https://github.com/dpzwick/ppiclF-doc/commit/e3f8d29173c14ee3fb0702b4303c9ceaa1e72ab7) so it is updated now.
