### 2019-04-29 #1400 @iljah
Thanks, I was asking about solved particles per second per core not number of particles.

What fraction of CPU time was used by the fluid solver and was that included in timings provided in performance section? I can imagine with only 300 particles per process the fluid solution would take much more time than particles which would improve particle scaling since time spent transferring particle data would be masked by fluid computation.
