### 2019-04-29 #1400 @dpzwick
In the scaling cases the fluid was not timed. Since Nek5000 was used, the times for the fluid solver are highly dependent on problem set setup since the fluid solver will iterate on a pressure poisson system. For this reason, the times on the scaling plots are only for ppiclF and not for any part of the fluid solver that it was coupled to.
