### 2019-04-30 #1400 @dpzwick
There were barriers set so that just the timings for ppiclF could be tested. In real applications there may be some overlap of communication with computation though.

Thanks for your insight and suggestions @iljah and @markadams!
