### 2019-05-02 #1400 @jedbrown
@dpzwick Some minor tweaks to the paper:
1. No need to cite the repository for this software; it appears in the margin.
2. Please check `[-@` and `@` syntax for your references. https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax
3. The displayed equation is completely general, but presumably one would not use this when `Y` is exclusively mesh-based field data, for example. A couple sentences about what particle-in-cell methods are and how ppiclF interfaces with field-based solvers would be useful. This could cite traditional publications that would be relevant to a user interested in further details or applications.
4. It would be useful to mention some other PIC software and how ppiclF compares to them. Note that some (e.g., [geodynamics](http://www.underworldcode.org/)) use only a handful of particles per cell while others (like plasma applications) use many thousands of particles per cell. A library might be better in one regime than another. [Underworld](http://www.underworldcode.org/) and [XGC1](https://epsi.pppl.gov/computing/xgc-1) are familiar to me, and [there](https://picksc.idre.ucla.edu/software/software-production-codes/) [are](https://fbpic.github.io/) [others](http://warp.lbl.gov/) likely to turn up if a prospective user searches.
