### 2019-05-02 a841f83
Updated paper.md and paper.bib for JOSS submission.
