+++ paper.bib
@@ -1,17 +1,9 @@
 @PhDThesis{Zwick,
-  title	= {Scalable highly-resolved {Euler-Lagrange} multiphase flow simulation with applications to shock tubes.},
-  school = {University of Florida},
-  author = {Zwick, D.},
-  year = {2019},
-  type = {{Doctoral} dissertation}
-}
-
-@online{ppiclF,
-  author = {Zwick, D.},
-  title = {{ppiclF.}},
-  year = 2019,
-  url = {https://github.com/dpzwick/ppiclF},
-  urldate = {2019-04-21}
+    author = "{Zwick, D.}",
+    title = {Scalable highly-resolved {Euler-Lagrange} multiphase flow simulation with applications to shock tubes.},
+    school = {University of Florida},
+    type = {{Doctoral} dissertation},
+    year=2019
 }
 
 @online{ppiclf-doc,
@@ -36,4 +28,24 @@
   organization = {The Center for Compressible Multiphase Turbulence at the University of Florida},
   url = {https://github.com/UFCCMT/Nek5000},
   urldate = {2019-04-21}
+}
+
+@InProceedings{liao_1996,
+  author	= {Liao, W. K. and Ou, C. W. and Ranka, S.},
+  booktitle	= {Proceedings of International Conference on Parallel
+		  Processing},
+  title		= {Dynamic alignment and distribution of irregularly coupled
+		  data arrays for scalable parallelization of
+		  particle-in-cell problems},
+  year		= {1996},
+  pages		= {57-61},
+  doi		= {10.1109/IPPS.1996.508039},
+  month		= {April}
+}
+
+@online{ZoltanHomePage,
+   author = {Boman, E. and Devine, K. and Fisk, L. A. and Heaphy, R. and Hendrickson, B. and Leung, V. and Vaughan, C. and Catalyurek, U. and Bozdag, D. and Mitchell, W.},
+   title = {{Zoltan} home page},
+   url = {https://cs.sandia.gov/Zoltan},
+   year = {1999}}
 }
