+++ paper.md
@@ -1,41 +1,51 @@
 ---
 title: 'ppiclF: A Parallel Particle-In-Cell Library in Fortran'
-tags:
-  - particle
-  - particle-in-cell
-  - ODE integrator
-  - Euler-Lagrange
-  - discrete element method
 authors:
-  - name: David Zwick
-    orcid: 0000-0001-5329-6341
-    affiliation: 1
-affiliations:
- - name: Department of Mechanical & Aerospace Engineering, University of Florida
-   index: 1
-date: 19 April 2019
+- affiliation: 1
+  name: David Zwick
+  orcid: 0000-0001-5329-6341
+date: "19 April 2019"
+output: pdf_document
 bibliography: paper.bib
+tags:
+- particle
+- particle-in-cell
+- ODE integrator
+- Euler-Lagrange
+- discrete element method
+affiliations:
+- index: 1
+  name: Department of Mechanical & Aerospace Engineering, University of Florida
 ---
+
 
 # Statement of Need
 
-Particle-in-cell simulations are ubiquitous in science and engineering. Relevant applications range from explosively driven compressible multiphase gas-solid flow to atomistic modeling. At the core of these diverse applications, a system of differential equations is advanced in time. As a result, the system given by
+Particle-in-cell (PIC) simulations are ubiquitous in science and engineering. Relevant applications range from explosively driven compressible multiphase gas-solid flow to atomistic modeling. In typical PIC simulations, individual particles are tracked as they move through a surrounding mesh, or cells. Throughout a simulation, particle-mesh and particle-particle interactions occur frequently. At the core of the particle portion of these simulations is a system of differential equations describing the physical behavior of each particle. As a result, the system given by
 
 $$\dfrac{d \mathbf{Y}}{dt} = \dot{\mathbf{Y}}$$
 
-must be solved subject to appropriate initial conditions $\mathbf{Y} (t = 0)$ and right-hand-side forcing $\dot{\mathbf{Y}}$. However, computational constraints make this a daunting task when there are billions of equations, whose forcing may even depend on the states of other equations. Adding to this difficulty is the fact that the particles are often added as an afterthought when building particle-in-cell software. As a result, typical particle-in-cell implementations may initially perform below expectations.
+must be solved subject to appropriate initial conditions $\mathbf{Y} (t = 0)$ and right-hand-side forcing $\dot{\mathbf{Y}}$. For PIC applications, $\mathbf{Y}$ often includes particle coordinates and other evolving particle states, while $\dot{\mathbf{Y}}$ may be composed of appropriate force models.
+
+While initially the given system of equations appears straightforward, computational constraints make this a daunting task when there are billions of equations, whose forcing may even depend on the states of other equations. Adding to this difficulty is the fact that the particles are often added as an afterthought when building new PIC software. As a result, typical PIC implementations may initially perform below expectations.
+
+In order to remedy these challenges, a parallel dynamic load-balancing approach can be taken where cells and particles are distributed to different processors in memory based on some measure of overall computational workload [@liao_1996]. This can be accomplished through the use of modern libraries, such as Zoltan [@ZoltanHomePage], where the cell and particle data is redistributed in memory. In preexisting cell-only applications though, the cells may already be evenly balanced and need not be altered. In this case, it is of practical importance that particles can unobtrusively be added to traditional cell-only codes without altering preexisting data structures, while at the same time achieving high parallel performance.
 
 # Summary
 
[…]
