### 2019-05-02 #1400 @dpzwick
Thanks for the feedback @jedbrown .

I updated the PDF based on your suggestions.
