### 2019-05-02 #1400 @jedbrown
Thanks. The ICPP `booktitle` field needs case protection. For the Nek5000 and CMT-nek references, I would use "in-text" style, `... Navier-Stokes solver Nek5000 [-@nek5000] ...`.
