### 2019-05-02 7a7a271
Updated references for JOSS submission.
