+++ paper.bib
@@ -32,8 +32,8 @@
 
 @InProceedings{liao_1996,
   author	= {Liao, W. K. and Ou, C. W. and Ranka, S.},
-  booktitle	= {Proceedings of International Conference on Parallel
-		  Processing},
+  booktitle	= {{Proceedings of International Conference on Parallel
+		  Processing}},
   title		= {Dynamic alignment and distribution of irregularly coupled
 		  data arrays for scalable parallelization of
 		  particle-in-cell problems},
