### 2019-05-02 #1400 @jedbrown
Ah, that formatting is because the bib entries don't have an `author` field. Surely there are some people, or "Nek5000 Development Team", that make sense to list as authors.
