### 2019-05-02 #1400 @dpzwick
I can probably find some people but that is the recommended citation style
on their website. What do you think?
