### 2019-05-03 #1400 @jedbrown
Oh, I see what you mean. That's unfortunate. I don't think there's a way to format the link optimally, so would type it manually `Nek5000 (2019)` and drop in the `nocite`. If that looks right, we'll be ready to tag and archive.
