### 2019-05-03 #1400 @jedbrown
Looks good to me. Please tag a release and archive on Zenodo or similar. We would prefer that both your source and documentation repository be archived together at the same DOI. I believe you can do that by directly uploading the files (instead of using the GitHub integration) from both repositories, then adding your metadata. Let me know if you have questions about that process, and please report the DOI back here when done.
