### 2019-05-04 #1400 @dpzwick
I uploaded to Zenodo with the source and documentation. The DOI is: 10.5281/zenodo.2667285.
