### 2019-05-04 #1400 @danielskatz
This DOI doesn't yet resolve for me - once it does, I will close this issue, and the publication process will be complete
