### 2019-05-05 #1400 @arfon
@kyleniemeyer @danielskatz - looks like the Crossref deposit API is down (http://status.crossref.org/incidents/psyzb7txpwbq) - suggest not accepting any papers until this is resolved.
