### 2019-05-05 #1400 @arfon
> @arfon it looks as though that issue was only temporary; can we try depositing again? I imagine this would have be done manually.

I tried already. Unfortunately Crossref are *actually* having major issues with their deposit API, they just don't seem to have posted publicly about it. Perhaps no-body is on call over the weekend :-\ ?
