### 2019-05-06 #1400 @arfon
Looks like the DOI is resolving now.
